# docs/swagger.yaml
definitions:
...
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: string
    type: object
...
  images.SetRandomSourceAlbumRequest:
    properties:
      album_id:
        type: integer
      include_all_public:
        type: boolean
    type: object
...
  key.req:
    properties:
      description:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/admin/random-source-album:
    get:
      consumes:
      - application/json
      description: Get the currently configured random image source album ID and whether
        to include all public images
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Get random source album configuration
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Configure the source album for random image selection. Use album_id=0
        for all public images, or specify an album ID
      parameters:
      - description: Random source configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/images.SetRandomSourceAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Configuration updated successfully
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Set random source album configuration
      tags:
      - admin
...
  /api/v1/images/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a single image file with optional storage strategy and visibility
        settings
      parameters:
      - description: Image file to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Storage strategy ID
        in: formData
        name: strategy_id
        type: string
      - description: 'Whether the image is public (default: true)'
        in: formData
        name: is_public
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Upload successful
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid form data or too many files
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "413":
          description: File too large
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Upload single image
      tags:
      - images
  /api/v1/images/uploads:
    post:
      consumes:
      - multipart/form-data
      description: Upload up to 10 images in a single batch request
      parameters:
      - collectionFormat: csv
        description: Image files to upload (max 10)
        in: formData
        items:
          type: file
        name: files
        required: true
        type: array
      - description: Storage strategy ID
        in: formData
        name: strategy_id
        type: string
      - description: 'Whether images are public (default: true)'
        in: formData
        name: is_public
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Batch upload completed
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid form data or too many files
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "413":
          description: Total file size too large
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Upload multiple images
      tags:
      - images
  /api/v1/token:
    get:
      consumes:
      - application/json
      description: Get all API keys for the authenticated user with their status and
        metadata
      produces:
      - application/json
      responses:
        "200":
          description: API key list
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: List API keys
      tags:
      - keys
    post:
      consumes:
      - application/json
      description: Create a new API key for programmatic access
      parameters:
      - description: API key creation request (description is optional)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/key.req'
      produces:
      - application/json
      responses:
        "200":
          description: API key created successfully
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Create API key
      tags:
      - keys
...
  /images/random:
    get:
      consumes:
      - application/json
      description: Get a random image, optionally filtered by album and dimensions
      parameters:
      - description: 'Response format: json or image (default: image)'
        in: query
        name: format
        type: string
      - description: Filter by album ID (0 = all public images, >0 = specific album)
        in: query
        name: album_id
        type: integer
      - description: Minimum image width
        in: query
        name: min_width
        type: integer
      - description: Minimum image height
        in: query
        name: min_height
        type: integer
      - description: Maximum image width
        in: query
        name: max_width
        type: integer
      - description: Maximum image height
        in: query
        name: max_height
        type: integer
      produces:
      - image/*
      - application/json
      responses:
        "200":
          description: Image metadata (when format=json)
          schema:
            $ref: '#/definitions/common.Response'
        "204":
          description: No content - no matching images found
          schema:
            type: string
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      summary: Get random image
      tags:
      - images
...
swagger: "2.0"